+++ ChangeLog.md
@@ -1,3 +1,7 @@
 # Changelog for HNumeric
 
-## Unreleased changes
+## 0.4.0.0
+
+* Add HNum.F (Functional Programming Library for HNum)
+* Add Convertable Class (HNumObj String -> HNumObj (Int or Integer or Double))
+* Add (!) for Vector (v ! 1 == (toList v) !! 1)